backend/requirements.txt
# FastAPI and related dependencies
fastapi==0.110.2
uvicorn[standard]==0.29.0

# Database dependencies
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pgvector==0.2.5  # PostgreSQL extension for vector search

# Embedding and LLM dependencies
numpy==1.26.4
faiss-cpu==1.7.4
langchain==0.1.15
vllm==0.3.2  # For serving LLaMA-3 API

# Other utilities
requests==2.31.0
python-dotenv==1.0.1
pytest==8.1.1  # For running tests
httpx==0.27.0  # For API testing in FastAPI